fix: Make AllUsers component check for mutual friends with new API data
Data returned from the API no longer included
who users were friends with and friends of. Instead,
to check if a user is a friend the list of all usernames
is compared to the list of the logged in users
friendsOf. If JohnDoe is a friendOf JaneDoe, then
JaneDoe's UserCard won't render with an 'Add Friend'
Button