refactor: Remove dnd-kit and revert to HTML drag and drop
dnd-kit still had limitations on drag and drop for
mobile devices. To simplify the codebase I've made
the decision to remove dnd-kit and the added complexity
that comes with setting it up in favor of the simplicity
of HTML's in-built drag and drop. Will still have to
come up with a solution for mobile users to add users
to Group Chats