refactor: Make Conversation use single conversation data instead of all
Conversation component now retrieves the conversation
data for a specific conversation using the recipients
name. Previously all conversations were achieved and
a great deal of filtering was done on the front end
to find the conversation between the user and the
recipient. As the Conversation component needs to
include future functionality to create group conversations
the method in which it retrieves the conversation required
needed to change.